update watchlist page - remove clerk auth
