+++ src/app/profile/page.tsx
@@ -2,7 +2,6 @@ import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
 
 const ProfilePage = () => {
-  // if no session, redirect to login
   const cookieStore = cookies();
   const session = cookieStore.get("session");
   if (!session) {
